+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Add documentation for `@__slot__` and `$` shorthand `@text` syntax [#37]
+- Fix broken `@cm_component` macro caused by changes to `Revise` APIs [#37]
+- Fix `StreamingRender` iterator on Julia 1.12+ [#37]
+
 ## [v2.2.2] - 2025-01-08
 
 ### Fixed
@@ -133,3 +139,4 @@ Initial release.
 [#30]: https://github.com/MichaelHatherly/HypertextTemplates.jl/issues/30
 [#34]: https://github.com/MichaelHatherly/HypertextTemplates.jl/issues/34
 [#35]: https://github.com/MichaelHatherly/HypertextTemplates.jl/issues/35
+[#37]: https://github.com/MichaelHatherly/HypertextTemplates.jl/issues/37